+++ src/CoordinateDependent.java
@@ -2,7 +2,7 @@
  * Defines an interface for Objects which are dependendt on coordinates to work correctly.
  */
 public interface CoordinateDependent {
-    void setCoordinates(double x, double y);
+    void setCoordinates(int x, int y);
     double getX();
     double getY();
 }
